requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
PyJWT==2.4.0
PyMySQL==1.0.2
python3-openid==3.2.0
requests-oauthlib==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.3.0
toml==0.10.2
uritemplate==4.1.1
